pyproject.toml
[tool.poetry]
name = "netbox-plugin-netisp"
version = "0.1.0.1"
description = "A plugin for ISPs to manage wireless and fiber operations using Netbox"
homepage = "https://github.com/MajesticFalcon/netbox-plugin-netisp"
repository = "https://github.com/MajesticFalcon/netbox-plugin-netisp"

authors = ["DB2 Team"]

packages = [
    { include = "netbox_netisp" },
]
[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^20.8b1"
invoke = "^1.5.0"
pylint = "^2.6.0"
pylint-django = "^2.4.2"
pydocstyle = "^5.1.1"
yamllint = "^1.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
